Move RegistrationDetails from registry_models to common
